Move sentry to src/utils
